requirements.txt
streamlit
google-generativeai
requests
beautifulsoup4
python-dotenv
python-docx
selenium
webdriver-manager
pillow